tiresias-api/pyproject.toml
[project]
name = "tiresias-api"
version = "1.0.0"
description = "AI-Powered Audio Description System API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "python-multipart>=0.0.6",
    "yt-dlp>=2024.1.0",
    "scenedetect[opencv]>=0.6.3",
    "opencv-python>=4.9.0",
    "ffmpeg-python>=0.2.0",
    "google-generativeai>=0.3.0",
    "anthropic>=0.8.0",
    "openai>=1.10.0",
    "pydub>=0.25.1",
    "librosa>=0.10.0",
    "soundfile>=0.12.1",
    "openai-whisper>=20231117",
    "elevenlabs>=0.2.0",
    "boto3>=1.34.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.2.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
